Huge Config Entites

+++ PMA.Sop.Domain/Product/Entities/Category.cs
@@ -1,17 +1,14 @@
 ﻿using System.Collections.Generic;
 using PMA.Sop.Domain.Common;
-using PMA.Sop.Framework.Domain;
 
 namespace PMA.Sop.Domain.Product.Entities
 {
     public class Category : BaseUserEntity<long>
     {
         public string Title { get; set; }
-
         public long? ParentId { get; set; }
         public string Icon { get; set; }
-
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
 
         #region Relationships
         public ICollection<Category> Children { get; set; }
